+++ examples/bulk.rs
@@ -20,4 +20,19 @@ fn main() {
     found = envmnt::is_all_exists(&vec!["ENV_VAR1", "ENV_VAR2"]);
 
     println!("All Found: {}", &found);
+
+    env = IndexMap::new();
+    env.insert("ENV_VAR1".to_string(), "MY VALUE".to_string());
+    env.insert("ENV_VAR2".to_string(), "MY VALUE2".to_string());
+
+    let eval_env = |value: String| {
+        let mut buffer = String::from("VALUE-");
+        buffer.push_str(&value);
+        buffer
+    };
+
+    envmnt::evaluate_and_set_all(&env, eval_env);
+
+    let value = envmnt::get_or_panic("ENV_VAR1");
+    println!("Value Is: {}", &value);
 }