+++ src/bulk.rs
@@ -17,6 +17,17 @@ pub(crate) fn set_all(env: &IndexMap<String, String>) {
     }
 }
 
+pub(crate) fn evaluate_and_set_all<F>(env: &IndexMap<String, String>, evaluate: F)
+where
+    F: Fn(String) -> String,
+{
+    for (key, value) in env.iter() {
+        let evaluated_value = evaluate(value.to_string());
+
+        environment::set(key.to_string(), evaluated_value);
+    }
+}
+
 pub(crate) fn is_any_exists<K: AsRef<OsStr>>(keys: &Vec<K>) -> bool {
     let mut found = false;